third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-valid-hsl-expected.txt

This is a testharness.js-based test.
Found 5 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] e.style['color'] = "hsl(30 -50% 60)" should set the property value
  assert_equals: serialization should be canonical expected "rgb(153, 153, 153)" but got "rgb(102, 153, 204)"
[FAIL] e.style['color'] = "hsl(0 -50 40%)" should set the property value
  assert_equals: serialization should be canonical expected "rgb(102, 102, 102)" but got "rgb(51, 153, 153)"
[FAIL] e.style['color'] = "hsl(30 -50 60)" should set the property value
  assert_equals: serialization should be canonical expected "rgb(153, 153, 153)" but got "rgb(102, 153, 204)"
[FAIL] e.style['color'] = "hsla(calc(50deg + (sign(1em - 10px) * 10deg)) -100 300 / 0.5)" should set the property value
  assert_equals: serialization should be canonical expected "hsl(calc(50deg + (10deg * sign(1em - 10px))) 0 300 / 0.5)" but got "hsl(calc(50deg + (10deg * sign(1em - 10px))) 0 255 / 0.5)"
[FAIL] e.style['color'] = "hsla(calc(50deg + (sign(1em - 10px) * 10deg)) -100% 300% / 0.5)" should set the property value
  assert_equals: serialization should be canonical expected "hsl(calc(50deg + (10deg * sign(1em - 10px))) 0 300 / 0.5)" but got "hsl(calc(50deg + (10deg * sign(1em - 10px))) 0 3 / 0.5)"
Harness: the test ran to completion.
